bug fix
1. transfer function map from (0,1024) to (0, 1023)
2. Vector3i scale operation (float) to (int)
3. add file existence check for MPI_File_open
4. delete multiple dataset option, # input para: 5 -> 4
5. ds.index_start 0->1

+++ Consts.h
@@ -68,14 +68,14 @@ public:
     Vector3i operator-(Vector3i const& rhs) const { Vector3i t(*this); t-=rhs; return t; }
     Vector3i operator*(Vector3i const& rhs) const { Vector3i t(*this); t*=rhs; return t; }
     Vector3i operator/(Vector3i const& rhs) const { Vector3i t(*this); t/=rhs; return t; }
-    Vector3i operator*(float scale) const { Vector3i t(*this); t*=scale; return t; }
-    Vector3i operator/(float scale) const { Vector3i t(*this); t/=scale; return t; }
+    Vector3i operator*(int scale) const { Vector3i t(*this); t*=scale; return t; }
+    Vector3i operator/(int scale) const { Vector3i t(*this); t/=scale; return t; }
     Vector3i& operator+=(Vector3i const& rhs) { x+=rhs.x, y+=rhs.y, z+=rhs.z; return *this; }
     Vector3i& operator-=(Vector3i const& rhs) { x-=rhs.x, y-=rhs.y, z-=rhs.z; return *this; }
     Vector3i& operator*=(Vector3i const& rhs) { x*=rhs.x, y*=rhs.y, z*=rhs.z; return *this; }
     Vector3i& operator/=(Vector3i const& rhs) { x/=rhs.x, y/=rhs.y, z/=rhs.z; return *this; }
-    Vector3i& operator*=(float scale) { x*=scale, y*=scale, z*=scale; return *this; }
-    Vector3i& operator/=(float scale) { x/=scale, y/=scale, z/=scale; return *this; }
+    Vector3i& operator*=(int scale) { x*=scale, y*=scale, z*=scale; return *this; }
+    Vector3i& operator/=(int scale) { x/=scale, y/=scale, z/=scale; return *this; }
 }; typedef Vector3i DataPoint;
 
 class Vector3f {
@@ -111,13 +111,6 @@ public:
         } else return false;
     }
 };    // start ---id---> end @ centroid
-
-//struct Edge {
-//    int id;
-//    int start;
-//    int end;
-//    Vector3i centroid;
-//};    // start ---id---> end @ centroid
 
 struct DataSet {
     int     index_start;
